Fix examples building
